requirements.txt
rubik-solver==0.2.0
pyserial==3.5
tkinter
